chore: configure runner node taints
We add node *taints*:

node_type=runner:NoSchedule
node_type=runner:NoExecute

This will ensure that only pods with the corresponding *tolerations* cab be scheduled and executed on these nodes. We will later add these tolerations to the pod specs.

We also add hint labels for cluster autoscaler so that it takes into accounts when deciding which nodes to scale.